+++ farben/src/core.rs
@@ -1,0 +1,5 @@
+//! Internals that the user may use for certain functions.
+//!
+//! This module does not expose anything, just items that the developer may use.
+
+pub use farben_core::ansi::style_to_ansi;